Rename spin package to configstack
