+++ configstack/test_helpers.go
@@ -1,4 +1,4 @@
-package spin
+package configstack
 
 import (
 	"github.com/gruntwork-io/terragrunt/errors"
@@ -111,7 +111,7 @@ func assertRunningModulesEqual(t *testing.T, expected *runningModule, actual *ru
 
 // We can't do a simple IsError comparison for UnrecognizedDependency because that error is a struct that
 // contains an array, and in Go, trying to compare arrays gives a "comparing uncomparable type
-// spin.UnrecognizedDependency" panic. Therefore, we have to compare that error more manually.
+// configstack.UnrecognizedDependency" panic. Therefore, we have to compare that error more manually.
 func assertErrorsEqual(t *testing.T, expected error, actual error, messageAndArgs ...interface{}) {
 	actual = errors.Unwrap(actual)
 	if expectedUnrecognized, isUnrecognizedDependencyError := expected.(UnrecognizedDependency); isUnrecognizedDependencyError {
